chore: Remove cjs support
