+++ client/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "../dist/client/index.js"
+}
