+++ package.json
@@ -8,28 +8,25 @@
   "sideEffects": false,
   "exports": {
     ".": {
-      "types": {
-        "require": "./dist/index.d.cts",
-        "import": "./dist/index.d.ts"
-      },
-      "require": "./dist/index.cjs",
-      "import": "./dist/index.mjs"
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
     },
     "./client": {
-      "types": {
-        "require": "./dist/client/index.d.cts",
-        "import": "./dist/client/index.d.ts"
-      },
-      "require": "./dist/client/index.cjs",
-      "import": "./dist/client/index.mjs"
+      "types": "./dist/client/index.d.ts",
+      "import": "./dist/client/index.js"
+    },
+    "./server": {
+      "types": "./dist/server/index.d.ts",
+      "import": "./dist/server/index.js"
     }
   },
-  "main": "./dist/index.mjs",
-  "module": "./dist/index.mjs",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "files": [
     "dist",
-    "client.d.ts"
+    "server",
+    "client"
   ],
   "scripts": {
     "dev": "tsup --watch --onSuccess \"pnpm --filter playground dev\"",
