+++ server/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "../dist/server/index.js"
+}
